+++ vidsjs/migrations/20160629154533-remove_connectsession.js
@@ -1,0 +1,32 @@
+'use strict';
+
+module.exports = {
+  up: function (queryInterface, Sequelize) {
+    return queryInterface.dropTable('ConnectSessions');
+  },
+
+  down: function (queryInterface, Sequelize) {
+    return queryInterface.createTable('ConnectSessions', {
+      id: {
+        allowNull: false,
+        autoIncrement: true,
+        primaryKey: true,
+        type: Sequelize.INTEGER
+      },
+      sid: {
+        type: Sequelize.STRING
+      },
+      expires: {
+        type: Sequelize.DATE
+      },
+      createdAt: {
+        allowNull: false,
+        type: Sequelize.DATE
+      },
+      updatedAt: {
+        allowNull: false,
+        type: Sequelize.DATE
+      }
+    });
+  }
+};